+++ src/lib.rs
@@ -414,9 +414,10 @@ macro_rules! fixed {
             $s_nbits_m1:expr, $s_nbits_m2:expr, $s_nbits_m3:expr, $s_nbits_m4:expr
         ),
         $nbytes:expr, $bytes_val:expr, $rev_bytes_val:expr, $be_bytes:expr, $le_bytes:expr,
-        $IFixed:ident, $UFixed:ident, $UInner:ident, $Signedness:tt,
+        $IFixed:ident, $UFixed:ident, $IInner:ident, $UInner:ident, $Signedness:tt,
         $LeEqU_C0:tt, $LeEqU_C1:tt, $LeEqU_C2:tt, $LeEqU_C3:tt,
-        $HasDouble:tt, $Double:ident, $DoubleInner:ty, $s_nbits_2:expr
+        $HasDouble:tt, $s_nbits_2:expr,
+        $Double:ident, $DoubleInner:ty, $IDouble:ident, $IDoubleInner:ty
     ) => {
         fixed! {
             $description,
@@ -425,9 +426,11 @@ macro_rules! fixed {
                 $s_nbits_m1, $s_nbits_m2, $s_nbits_m3, $s_nbits_m4
             ),
             $nbytes, $bytes_val, $rev_bytes_val, $be_bytes, $le_bytes,
-            $IFixed[stringify!($IFixed)], $UFixed[stringify!($UFixed)], $UInner, $Signedness,
+            $IFixed[stringify!($IFixed)], $UFixed[stringify!($UFixed)],
+            $IInner, $UInner, $Signedness,
             $LeEqU_C0, $LeEqU_C1, $LeEqU_C2, $LeEqU_C3,
-            $HasDouble, $Double[stringify!($Double)], $DoubleInner, $s_nbits_2
+            $HasDouble, $s_nbits_2,
+            $Double[stringify!($Double)], $DoubleInner, $IDouble, $IDoubleInner
         }
     };
     (
@@ -437,9 +440,11 @@ macro_rules! fixed {
             $s_nbits_m1:expr, $s_nbits_m2:expr, $s_nbits_m3:expr, $s_nbits_m4:expr
         ),
         $nbytes:expr, $bytes_val:expr, $rev_bytes_val:expr, $be_bytes:expr, $le_bytes:expr,
-        $IFixed:ident[$s_ifixed:expr], $UFixed:ident[$s_ufixed:expr], $UInner:ident, $Signedness:tt,
+        $IFixed:ident[$s_ifixed:expr], $UFixed:ident[$s_ufixed:expr],
+        $IInner:ident, $UInner:ident, $Signedness:tt,
         $LeEqU_C0:tt, $LeEqU_C1:tt, $LeEqU_C2:tt, $LeEqU_C3:tt,
-        $HasDouble:tt, $Double:ident[$s_double:expr], $DoubleInner:ty, $s_nbits_2:expr
+        $HasDouble:tt, $s_nbits_2:expr,
+        $Double:ident[$s_double:expr], $DoubleInner:ty, $IDouble:ident, $IDoubleInner:ty
     ) => {
         comment! {
             $description, "-bit ",
@@ -549,8 +554,10 @@ assert_eq!(two_point_75.to_string(), \"2.8\");
         fixed_no_frac! {
             $Fixed[$s_fixed]($Inner[$s_inner], $LeEqU, $UNbits, $s_nbits, $s_nbits_m1, $s_nbits_m2),
             $nbytes, $bytes_val, $rev_bytes_val, $be_bytes, $le_bytes,
-            $IFixed[$s_ifixed], $UFixed[$s_ufixed], $UInner, $Signedness,
-            $HasDouble, $Double[$s_double], $DoubleInner, $s_nbits_2
+            $IFixed[$s_ifixed], $UFixed[$s_ufixed],
+            $IInner, $UInner, $Signedness,
+            $HasDouble, $s_nbits_2,
+            $Double[$s_double], $DoubleInner, $IDouble, $IDoubleInner
         }
         // inherent methods that require Frac bounds, and cannot be const
         fixed_frac! {
@@ -569,25 +576,25 @@ fixed! {
     "An eight",
     FixedU8(u8, LeEqU8, U8, "8", "7", "6", "5", "4"),
     1, "0x12", "0x12", "[0x12]", "[0x12]",
-    FixedI8, FixedU8, u8, Unsigned,
+    FixedI8, FixedU8, i8, u8, Unsigned,
     U8, U7, U6, U5,
-    True, FixedU16, u16, "16"
+    True, "16", FixedU16, u16, FixedI16, i16
 }
 fixed! {
     "A 16",
     FixedU16(u16, LeEqU16, U16, "16", "15", "14", "13", "12"),
     2, "0x1234", "0x3412", "[0x12, 0x34]", "[0x34, 0x12]",
-    FixedI16, FixedU16, u16, Unsigned,
+    FixedI16, FixedU16, i16, u16, Unsigned,
     U16, U15, U14, U13,
-    True, FixedU32, u32, "32"
+    True, "32", FixedU32, u32, FixedI32, i32
 }
 fixed! {
     "A 32",
     FixedU32(u32, LeEqU32, U32, "32", "31", "30", "29", "28"),
     4, "0x1234_5678", "0x7856_3412", "[0x12, 0x34, 0x56, 0x78]", "[0x78, 0x56, 0x34, 0x12]",
-    FixedI32, FixedU32, u32, Unsigned,
+    FixedI32, FixedU32, i32, u32, Unsigned,
     U32, U31, U30, U29,
-    True, FixedU64, u64, "64"
+    True, "64", FixedU64, u64, FixedI64, i64
 }
 fixed! {
     "A 64",
@@ -595,9 +602,9 @@ fixed! {
     8, "0x1234_5678_9ABC_DE0F", "0x0FDE_BC9A_7856_3412",
     "[0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0x0F]",
     "[0x0F, 0xDE, 0xBC, 0x9A, 0x78, 0x56, 0x34, 0x12]",
-    FixedI64, FixedU64, u64, Unsigned,
+    FixedI64, FixedU64, i64, u64, Unsigned,
     U64, U63, U62, U61,
-    True, FixedU128, u128, "128"
+    True, "128", FixedU128, u128, FixedI128, i128
 }
 fixed! {
     "A 128",
@@ -608,33 +615,33 @@ fixed! {
      0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08]",
     "[0x08, 0x07, 0x06, 0x05, 0x04, 0x03, 0x02, 0x01, \
      0xF0, 0xDE, 0xBC, 0x9A, 0x78, 0x56, 0x34, 0x12]",
-    FixedI128, FixedU128, u128, Unsigned,
+    FixedI128, FixedU128, i128, u128, Unsigned,
     U128, U127, U126, U125,
-    False, FixedU128, u128, "128"
+    False, "128", FixedU128, u128, FixedI128, i128
 }
 fixed! {
     "An eight",
     FixedI8(i8, LeEqU8, U8, "8", "7", "6", "5", "4"),
     1, "0x12", "0x12", "[0x12]", "[0x12]",
-    FixedI8, FixedU8, u8, Signed,
+    FixedI8, FixedU8, i8, u8, Signed,
     U7, U6, U5, U4,
-    True, FixedI16, i16, "16"
+    True, "16", FixedI16, i16, FixedI16, i16
 }
 fixed! {
     "A 16",
     FixedI16(i16, LeEqU16, U16,"16", "15", "14", "13", "12"),
     2, "0x1234", "0x3412", "[0x12, 0x34]", "[0x34, 0x12]",
-    FixedI16, FixedU16, u16, Signed,
+    FixedI16, FixedU16, i16, u16, Signed,
     U15, U14, U13, U12,
-    True, FixedI32, i32, "32"
+    True, "32", FixedI32, i32, FixedI32, i32
 }
 fixed! {
     "A 32",
     FixedI32(i32, LeEqU32, U32, "32", "31", "30", "29", "28"),
     4, "0x1234_5678", "0x7856_3412", "[0x12, 0x34, 0x56, 0x78]", "[0x78, 0x56, 0x34, 0x12]",
-    FixedI32, FixedU32, u32, Signed,
+    FixedI32, FixedU32, i32, u32, Signed,
     U31, U30, U29, U28,
-    True, FixedI64, i64, "64"
+    True, "64", FixedI64, i64, FixedI64, i64
 }
 fixed! {
     "A 64",
@@ -642,9 +649,9 @@ fixed! {
     8, "0x1234_5678_9ABC_DE0F", "0x0FDE_BC9A_7856_3412",
     "[0x12, 0x34, 0x56, 0x78, 0x9A, 0xBC, 0xDE, 0x0F]",
     "[0x0F, 0xDE, 0xBC, 0x9A, 0x78, 0x56, 0x34, 0x12]",
-    FixedI64, FixedU64, u64, Signed,
+    FixedI64, FixedU64, i64, u64, Signed,
     U63, U62, U61, U60,
-    True, FixedI128, i128, "128"
+    True, "128", FixedI128, i128, FixedI128, i128
 }
 fixed! {
     "A 128",
@@ -655,9 +662,9 @@ fixed! {
      0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08]",
     "[0x08, 0x07, 0x06, 0x05, 0x04, 0x03, 0x02, 0x01, \
      0xF0, 0xDE, 0xBC, 0x9A, 0x78, 0x56, 0x34, 0x12]",
-    FixedI128, FixedU128, u128, Signed,
+    FixedI128, FixedU128, i128, u128, Signed,
     U127, U126, U125, U124,
-    False, FixedI128, i128, "128"
+    False, "128", FixedI128, i128, FixedI128, i128
 }
 
 /// The bit representation of a *binary128* floating-point number (`f128`).
@@ -1432,5 +1439,134 @@ mod tests {
             I1F31::from_num(-1).overflowing_recip(),
             (I1F31::from_num(-1), false)
         );
+    }
+
+    #[test]
+    fn wide_mul_mixed() {
+        // +7FFF.FFFF * 7FFF.FFFF = +3FFF_FFFE.0000_0001
+        // 7FFF.FFFF * 7FFF.FFFF = 3FFF_FFFE.0000_0001
+        // +7FFF.FFFF * +7FFF.FFFF = +3FFF_FFFE.0000_0001
+        let s = I16F16::MAX;
+        let u = U16F16::MAX >> 1u32;
+        let t = U16F16::from_bits(s.to_bits() as u32);
+        let v = I16F16::from_bits(u.to_bits() as i32);
+        assert_eq!(s.wide_mul_unsigned(u).to_bits(), 0x3FFF_FFFF_0000_0001);
+        assert_eq!(t.wide_mul(u).to_bits(), 0x3FFF_FFFF_0000_0001);
+        assert_eq!(s.wide_mul(v).to_bits(), 0x3FFF_FFFF_0000_0001);
+        assert_eq!(s.wide_mul_unsigned(u), u.wide_mul_signed(s));
+        assert_eq!(t.wide_mul(u), u.wide_mul(t));
+        assert_eq!(s.wide_mul(v), v.wide_mul(s));
+
+        // +7FFF.FFFF * 8000.0000 = +3FFF_FFFF.8000_0000
+        // 7FFF.FFFF * 8000.0000 = 3FFF_FFFF.8000_0000
+        // +7FFF.FFFF * -8000.0000 = -3FFF_FFFF.8000_0000
+        let s = I16F16::MAX;
+        let u = !(U16F16::MAX >> 1u32);
+        let t = U16F16::from_bits(s.to_bits() as u32);
+        let v = I16F16::from_bits(u.to_bits() as i32);
+        assert_eq!(s.wide_mul_unsigned(u).to_bits(), 0x3FFF_FFFF_8000_0000);
+        assert_eq!(t.wide_mul(u).to_bits(), 0x3FFF_FFFF_8000_0000);
+        assert_eq!(s.wide_mul(v).to_bits(), -0x3FFF_FFFF_8000_0000);
+        assert_eq!(s.wide_mul_unsigned(u), u.wide_mul_signed(s));
+        assert_eq!(t.wide_mul(u), u.wide_mul(t));
+        assert_eq!(s.wide_mul(v), v.wide_mul(s));
+
+        // +7FFF.FFFF * FFFF.FFFF = +7FFF_FFFE.8000_0001
+        // 7FFF.FFFF * FFFF.FFFF = 7FFF_FFFE.8000_0001
+        // +7FFF.FFFF * -0000.0001 = -0000_0000.7FFF_FFFF
+        let s = I16F16::MAX;
+        let u = U16F16::MAX;
+        let t = U16F16::from_bits(s.to_bits() as u32);
+        let v = I16F16::from_bits(u.to_bits() as i32);
+        assert_eq!(s.wide_mul_unsigned(u).to_bits(), 0x7FFF_FFFE_8000_0001);
+        assert_eq!(t.wide_mul(u).to_bits(), 0x7FFF_FFFE_8000_0001);
+        assert_eq!(s.wide_mul(v).to_bits(), -0x0000_0000_7FFF_FFFF);
+        assert_eq!(s.wide_mul_unsigned(u), u.wide_mul_signed(s));
+        assert_eq!(t.wide_mul(u), u.wide_mul(t));
+        assert_eq!(s.wide_mul(v), v.wide_mul(s));
+
+        // -8000.0000 * 7FFF.FFFF = -3FFF_FFFF.8000_0000
+        // 8000.0000 * 7FFF.FFFF = 3FFF_FFFF.8000_0000
+        // -8000.0000 * +7FFF.FFFF = -3FFF_FFFF.8000_0000
+        let s = I16F16::MIN;
+        let u = U16F16::MAX >> 1u32;
+        let t = U16F16::from_bits(s.to_bits() as u32);
+        let v = I16F16::from_bits(u.to_bits() as i32);
+        assert_eq!(s.wide_mul_unsigned(u).to_bits(), -0x3FFF_FFFF_8000_0000);
+        assert_eq!(t.wide_mul(u).to_bits(), 0x3FFF_FFFF_8000_0000);
+        assert_eq!(s.wide_mul(v).to_bits(), -0x3FFF_FFFF_8000_0000);
+        assert_eq!(s.wide_mul_unsigned(u), u.wide_mul_signed(s));
+        assert_eq!(t.wide_mul(u), u.wide_mul(t));
+        assert_eq!(s.wide_mul(v), v.wide_mul(s));
+
+        // -8000.0000 * 8000.0000 = -4000_0000.0000_0000
+        // 8000.0000 * 8000.0000 = 4000_0000.0000_0000
+        // -8000.0000 * -8000.0000 = +4000_0000.0000_0000
+        let s = I16F16::MIN;
+        let u = !(U16F16::MAX >> 1u32);
+        let t = U16F16::from_bits(s.to_bits() as u32);
+        let v = I16F16::from_bits(u.to_bits() as i32);
+        assert_eq!(s.wide_mul_unsigned(u).to_bits(), -0x4000_0000_0000_0000);
+        assert_eq!(t.wide_mul(u).to_bits(), 0x4000_0000_0000_0000);
+        assert_eq!(s.wide_mul(v).to_bits(), 0x4000_0000_0000_0000);
+        assert_eq!(s.wide_mul_unsigned(u), u.wide_mul_signed(s));
+        assert_eq!(t.wide_mul(u), u.wide_mul(t));
+        assert_eq!(s.wide_mul(v), v.wide_mul(s));
+
+        // -8000.0000 * FFFF.FFFF = -7FFF_FFFF.8000_0000
+        // 8000.0000 * FFFF.FFFF = 7FFF_FFFF.8000_0000
+        // -8000.0000 * -0000.0001 = +0000_0000.8000_0000
+        let s = I16F16::MIN;
+        let u = U16F16::MAX;
+        let t = U16F16::from_bits(s.to_bits() as u32);
+        let v = I16F16::from_bits(u.to_bits() as i32);
+        assert_eq!(s.wide_mul_unsigned(u).to_bits(), -0x7FFF_FFFF_8000_0000);
+        assert_eq!(t.wide_mul(u).to_bits(), 0x7FFF_FFFF_8000_0000);
+        assert_eq!(s.wide_mul(v).to_bits(), 0x8000_0000);
+        assert_eq!(s.wide_mul_unsigned(u), u.wide_mul_signed(s));
+        assert_eq!(t.wide_mul(u), u.wide_mul(t));
+        assert_eq!(s.wide_mul(v), v.wide_mul(s));
+
+        // -0000.0001 * 7FFF.FFFF = -0000_0000.7FFF_FFFF
+        // FFFF.FFFF * 7FFF.FFFF = 7FFF_FFFE.8000_0001
+        // -0000.0001 * +7FFF.FFFF = -0000_0000.7FFF_FFFF
+        let s = -I16F16::DELTA;
+        let u = U16F16::MAX >> 1u32;
+        let t = U16F16::from_bits(s.to_bits() as u32);
+        let v = I16F16::from_bits(u.to_bits() as i32);
+        assert_eq!(s.wide_mul_unsigned(u).to_bits(), -0x0000_0000_7FFF_FFFF);
+        assert_eq!(t.wide_mul(u).to_bits(), 0x7FFF_FFFE_8000_0001);
+        assert_eq!(s.wide_mul(v).to_bits(), -0x0000_0000_7FFF_FFFF);
+        assert_eq!(s.wide_mul_unsigned(u), u.wide_mul_signed(s));
+        assert_eq!(t.wide_mul(u), u.wide_mul(t));
+        assert_eq!(s.wide_mul(v), v.wide_mul(s));
+
+        // -0000.0001 * 8000.0000 = -0000_0000.8000_0000
+        // FFFF.FFFF * 8000.0000 = 7FFF_FFFF.8000_0000
+        // -0000.0001 * -8000.0000 = +0000_0000.8000_0000
+        let s = -I16F16::DELTA;
+        let u = !(U16F16::MAX >> 1u32);
+        let t = U16F16::from_bits(s.to_bits() as u32);
+        let v = I16F16::from_bits(u.to_bits() as i32);
+        assert_eq!(s.wide_mul_unsigned(u).to_bits(), -0x0000_0000_8000_0000);
+        assert_eq!(t.wide_mul(u).to_bits(), 0x7FFF_FFFF_8000_0000);
+        assert_eq!(s.wide_mul(v).to_bits(), 0x0000_0000_8000_0000);
+        assert_eq!(s.wide_mul_unsigned(u), u.wide_mul_signed(s));
+        assert_eq!(t.wide_mul(u), u.wide_mul(t));
+        assert_eq!(s.wide_mul(v), v.wide_mul(s));
+
+        // -0000.0001 * FFFF.FFFF = -0000_0000.FFFF_FFFF
+        // FFFF.FFFF * FFFF.FFFF = FFFF_FFFE.0000_0001
+        // -0000.0001 * -0000.0001 = +0000_0000.0000_0001
+        let s = -I16F16::DELTA;
+        let u = U16F16::MAX;
+        let t = U16F16::from_bits(s.to_bits() as u32);
+        let v = I16F16::from_bits(u.to_bits() as i32);
+        assert_eq!(s.wide_mul_unsigned(u).to_bits(), -0x0000_0000_FFFF_FFFF);
+        assert_eq!(t.wide_mul(u).to_bits(), 0xFFFF_FFFE_0000_0001);
+        assert_eq!(s.wide_mul(v).to_bits(), 0x0000_0000_0000_0001);
+        assert_eq!(s.wide_mul_unsigned(u), u.wide_mul_signed(s));
+        assert_eq!(t.wide_mul(u), u.wide_mul(t));
+        assert_eq!(s.wide_mul(v), v.wide_mul(s));
     }
 }